mcpapp/requirements.txt
fastmcp==0.1.0
langgraph==0.2.34
langchain-community==0.2.16
python-dotenv==1.0.1
uvicorn==0.30.6
dashscope==1.20.9
langchain==0.2.16
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
aiohttp==3.10.10